requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple/
absl-py==0.13.0
argon2-cffi==21.1.0; python_version >= '3.5'
astunparse==1.6.3
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
bleach==4.1.0; python_version >= '3.6'
bunch==1.0.1
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
clang==5.0
cycler==0.10.0
debugpy==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
decorator==5.0.9; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints==0.3; python_version >= '2.7'
flatbuffers==1.12
gast==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==1.35.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0; python_version >= '3.6'
idna==3.2; python_version >= '3'
ipykernel==6.3.1; python_version >= '3.7'
ipython-genutils==0.2.0
ipython==7.16.3; python_version >= '3.3'
ipywidgets==7.6.4; python_version >= '3.6'
jedi==0.18.0; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
joblib==1.0.1; python_version >= '3.6'
jsonschema==3.2.0
jupyter-client==7.0.2; python_full_version >= '3.6.1'
jupyter-console==6.4.0; python_version >= '3.6'
jupyter-core==4.7.1; python_version >= '3.6'
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1; python_version >= '3.6'
keras-preprocessing==1.1.2
keras==2.6.0
kiwisolver==1.3.2; python_version >= '3.7'
markdown==3.3.4; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
matplotlib-inline==0.1.3; python_version >= '3.5'
matplotlib==3.2.1
mistune==0.8.4
nbclient==0.5.4; python_full_version >= '3.6.1'
nbconvert==6.1.0; python_version >= '3.7'
nbformat==5.1.3; python_version >= '3.5'
nest-asyncio==1.5.1; python_version >= '3.5'
notebook==6.4.3; python_version >= '3.6'
numpy==1.19.5; python_version >= '3.9'
oauthlib==3.1.1; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
packaging==21.0; python_version >= '3.6'
pandocfilters==1.4.3
parso==0.8.2; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==8.3.2
prometheus-client==0.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prompt-toolkit==3.0.20; python_full_version >= '3.6.2'
protobuf==3.17.3
ptyprocess==0.7.0; os_name != 'nt'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.10.0; python_version >= '3.5'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrsistent==0.18.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyzmq==22.2.1; python_version >= '3.6'
qtconsole==5.1.1; python_version >= '3.6'
qtpy==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rdkit-pypi==2021.3.4
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.6'
scikit-learn==0.24.2
scipy==1.7.1; python_version < '3.10' and python_version >= '3.7'
send2trash==1.8.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0; python_version >= '3.6'
tensorflow-estimator==2.6.0
tensorflow==2.6.0rc2
termcolor==1.1.0
terminado==0.12.1; python_version >= '3.6'
testpath==0.5.0; python_version >= '3.5'
threadpoolctl==2.2.0; python_version >= '3.6'
tornado==6.1; python_version >= '3.5'
tqdm==4.62.0
traitlets==5.1.0; python_version >= '3.7'
typing-extensions==3.7.4.3
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.1; python_version >= '3.6'
wheel==0.37.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
widgetsnbextension==3.5.1; python_version >= '3.6'
wrapt==1.12.1
